# apiDocumantatiob.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Видеохостинг API",
    "description": "API для платформы видеохостинга с поддержкой аутентификации, загрузки видео, комментариев, лайков/дизлайков и управления пользователями. Роли: администратор, менеджер, авторизованный клиент, гость.",
    "version": "1.0.0",
    "contact": {
      "name": "Разработчик",
      "email": "support@videohosting.ru"
    }
  },
  "servers": [
    {
      "url": "https://api.videohosting.ru/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/v1",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен, полученный при аутентификации"
      }
    },
    "schemas": {
...
      "Authorization": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "login": { "type": "string", "example": "user@example.com" },
          "id_user": { "type": "integer", "example": 1 },
          "id_role": { "type": "integer", "example": 2 },
          "id_avatar": { "type": "integer", "nullable": true, "example": 3 },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "role_name": { "type": "string", "example": "Менеджер" }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["login", "password"],
        "properties": {
          "login": { "type": "string", "example": "user@example.com" },
          "password": { "type": "string", "format": "password", "example": "secret123" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/Authorization" }
        }
      },
      "RegistrationRequest": {
        "type": "object",
        "required": ["login", "password", "firstName", "secondName", "dateOfBirth"],
        "properties": {
          "login": { "type": "string", "example": "newuser@example.com" },
          "password": { "type": "string", "format": "password", "example": "pass123" },
          "firstName": { "type": "string", "example": "Анна" },
          "secondName": { "type": "string", "example": "Петрова" },
          "fatherName": { "type": "string", "nullable": true, "example": "Сергеевна" },
          "dateOfBirth": { "type": "string", "format": "date", "example": "1995-06-15" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "description": "Стандартный объект ошибки"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Регистрация нового пользователя",
        "description": "Создаёт записи в таблицах users и authorizations, присваивает роль клиента (id_role = 3), аватар не назначается.",
        "operationId": "registerUser",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistrationRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Регистрация успешна",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Некорректные данные или пользователь уже существует",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Аутентификация пользователя",
        "description": "Проверяет логин/пароль, возвращает JWT токен и данные авторизации.",
        "operationId": "loginUser",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Неверный логин или пароль",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}